重构配方修改器
