+++ src/main/java/com/gto/gtocore/common/data/machines/MultiBlockC.java
@@ -679,7 +679,7 @@ public interface MultiBlockC {
             .tooltipsKey("gtocore.machine.automatic_chisel.tooltip.1")
             .tooltipsText("The higher the coil temperature, the faster the operation speed", "线圈温度越高，运行速度越快")
             .recipe(GTORecipeTypes.CHEMICAL_VAPOR_DEPOSITION_RECIPES)
-            .recipeModifiers((machine, r) -> recipe -> GTORecipeModifiers.recipeReduction(machine, recipe, 1, Math.log(900) / Math.log(((ICoilMachine) machine).gto$getTemperature())), GTORecipeModifiers.OVERCLOCKING)
+            .recipeModifiers((machine, recipe) -> GTORecipeModifiers.recipeReduction(machine, recipe, 1, Math.log(900) / Math.log(((ICoilMachine) machine).gto$getTemperature())), GTORecipeModifiers.OVERCLOCKING)
             .block(GTBlocks.CASING_PTFE_INERT)
             .pattern(definition -> FactoryBlockPattern.start(RelativeDirection.RIGHT, RelativeDirection.UP, RelativeDirection.BACK)
                     .aisle("     BBHBB", "     BGGGB", "     BGGGB", "     BGGGB", "     EEEEE")
@@ -708,7 +708,7 @@ public interface MultiBlockC {
             .tooltipsKey("gtocore.machine.automatic_chisel.tooltip.1")
             .tooltipsText("The higher the glass tier, the faster the operation speed", "玻璃等级越高，运行速度越快")
             .recipe(GTORecipeTypes.PHYSICAL_VAPOR_DEPOSITION_RECIPES)
-            .recipeModifiers((machine, r) -> recipe -> GTORecipeModifiers.recipeReduction(machine, recipe, 1, Math.sqrt(1.0D / ((ITierCasingMachine) machine).getCasingTier(GLASS_TIER))), GTORecipeModifiers.OVERCLOCKING)
+            .recipeModifiers((machine, recipe) -> GTORecipeModifiers.recipeReduction(machine, recipe, 1, Math.sqrt(1.0D / ((ITierCasingMachine) machine).getCasingTier(GLASS_TIER))), GTORecipeModifiers.OVERCLOCKING)
             .block(GTBlocks.PLASTCRETE)
             .pattern(definition -> FactoryBlockPattern.start()
                     .aisle("AAAA     ", "AAAA     ", "JJJA     ", "AAJA     ", "AAAA     ")
