+++ src/main/java/com/gto/gtocore/mixin/gtm/registry/GTMachineUtilsMixin.java
@@ -172,7 +172,7 @@ public final class GTMachineUtilsMixin {
                         .editableUI(SimpleGeneratorMachine.EDITABLE_UI_CREATOR.apply(GTCEu.id(name), recipeType))
                         .rotationState(RotationState.ALL)
                         .recipeType(recipeType)
-                        .recipeModifier(GTORecipeModifiers::simpleGeneratorMachineModifier)
+                        .recipeModifier(GTORecipeModifiers.SIMPLE_GENERATOR_MACHINEMODIFIER)
                         .addOutputLimit(ItemRecipeCapability.CAP, 0)
                         .addOutputLimit(FluidRecipeCapability.CAP, 0)
                         .renderer(() -> new SimpleGeneratorMachineRenderer(tier, GTCEu.id("block/generators/" + name)))
